+++ Octokit.Reactive/Clients/IObservableChecksClient.cs
@@ -2,7 +2,6 @@
 {
     public interface IObservableChecksClient
     {
-        IObservableCheckRunsClient Run { get; }
         IObservableCheckSuitesClient Suite { get; }
     }
 }
