+++ Octokit.Reactive/Clients/ObservableChecksClient.cs
@@ -4,11 +4,8 @@
     {
         public ObservableChecksClient(IGitHubClient gitHubClient)
         {
-            Run = new ObservableCheckRunsClient(gitHubClient);
             Suite = new ObservableCheckSuitesClient(gitHubClient);
         }
-
-        public IObservableCheckRunsClient Run { get; private set; }
 
         public IObservableCheckSuitesClient Suite { get; private set; }
     }
