+++ Octokit/Clients/IChecksClient.cs
@@ -2,8 +2,6 @@
 {
     public interface IChecksClient
     {
-        ICheckRunsClient Run { get; }
-
         ICheckSuitesClient Suite { get; }
     }
 }
