--- Octokit/Models/Response/ActivityPayloads/CheckRunEventPayload.cs
@@ -1,11 +1,0 @@
-﻿using System.Diagnostics;
-
-namespace Octokit
-{
-    [DebuggerDisplay("{DebuggerDisplay,nq}")]
-    public class CheckRunEventPayload : ActivityPayload
-    {
-        public string Action { get; protected set; }
-        public CheckRun CheckRun { get; protected set; }
-    }
-}
